+++ src/Credfeto.Dispatcher.GitHub/Models/NotificationSerializerContext.cs
@@ -12,6 +12,7 @@ namespace Credfeto.Dispatcher.GitHub.Models;
 [JsonSerializable(typeof(ApiIssue[]))]
 [JsonSerializable(typeof(ApiIssuePullRequest))]
 [JsonSerializable(typeof(ApiUserRepo[]))]
+[JsonSerializable(typeof(ApiRequiredStatusChecks))]
 [JsonSerializable(typeof(GraphQlRequest))]
 [JsonSerializable(typeof(GraphQlResponse))]
 [JsonSourceGenerationOptions(PropertyNamingPolicy = JsonKnownNamingPolicy.SnakeCaseLower)]